prefect/requirements.txt
google-cloud-storage==2.8.0
pandas==1.5.3
pandas-gbq==0.19.1
pandera==0.14.5
prefect==2.9.0
prefect-gcp==0.3.0
prefect_gcp[cloud_storage]
pyarrow==11.0.0
requests==2.28.2